Renamed files, changed to functional component, and removed Events.js
