--- src/components/Calendar/Calendar.js
@@ -1,28 +1,0 @@
-import React, { Component } from "react";
-import { Calendar, momentLocalizer } from "react-big-calendar";
-import moment from "moment";
-import events from "./Events";
-
-import "react-big-calendar/lib/css/react-big-calendar.css";
-
-const localizer = momentLocalizer(moment);
-
-export class BetterCalendar extends Component {
-  render() {
-    return (
-      <div>
-        <Calendar
-          localizer={localizer}
-          defaultDate={new Date()}
-          defaultView="month"
-          events={events}
-          style={{
-            height: "70vh",
-            width: "70vw",
-            margin: "0 auto"
-          }}
-        />
-      </div>
-    );
-  }
-}
