+++ src/pages/ProfilePage/ProfilePage.js
@@ -1,13 +1,13 @@
 import React from "react";
 
-import { BetterCalendar } from "../../components/Calendar/Calendar";
+import MeetingsCalendar from "../../components/MeetingsCalendar/MeetingsCalendar";
 
 export function ProfilePage() {
   return (
     <>
       <h1>Profile page</h1>
       <p>This is the profile page</p>
-      <BetterCalendar />
+      <MeetingsCalendar />
     </>
   );
 }
